Start to move Vmware Refresher and Parser, too
