+++ vmdb/app/models/ems/base_provider.rb
@@ -1,0 +1,6 @@
+module Ems
+class BaseProvider < ExtManagementSystem
+end
+end
+
+require_dependency 'ems/base_provider/refresher'
