+++ vmdb/app/models/ems/infra_provider.rb
@@ -1,5 +1,5 @@
 module Ems
-class InfraProvider < ExtManagementSystem
+class InfraProvider < BaseProvider
   ::EmsInfra = self # XXX
 
   SUBCLASSES = %w{
